agent/requirements.txt
# AWS Support Agent Dependencies

# AgentCore Runtime SDK (required for BedrockAgentCoreApp)
bedrock-agentcore

# Strands Agent Framework
strands-agents>=1.13.0

# Strands Tools
strands-agents-tools>=0.2.12

# Git operations
GitPython>=3.1.40

# AWS SDK
boto3>=1.35.0
botocore>=1.35.0

# Optional
typing-extensions>=4.8.0
